feat(config): add companion settings application to AppConfig and refactor init_chat_model
BREAKING CHANGE: init_chat_model function is now moved from main and streamlit_chat to AppConfig class, and MAX_TOKENS constant is now defined within the config module. Existing imports and usages of init_chat_model and MAX_TOKENS in main and streamlit_chat files should be updated to reflect these changes. Additionally, redundant application of companion settings has been removed from SlackAppConfig and StreamlitAppConfig.
