+++ config/app_config.py
@@ -1,10 +1,14 @@
 from __future__ import annotations
 
+import json
 from abc import ABC, abstractmethod
 from configparser import ConfigParser
 from typing import Any
 
 from google.cloud import firestore
+from langchain.chat_models import ChatOpenAI
+
+MAX_TOKENS = 1023
 
 
 def safely_get_field(
@@ -73,6 +77,56 @@ class AppConfig(ABC):
                 setting in self.config["firebase"]
             ), f"Missing configuration for {setting}"
 
+    def _apply_settings_from_companion(
+        self, companion: firestore.DocumentSnapshot
+    ) -> None:
+        """
+        Applies settings from the given companion Firestore document to the provided app configuration.
+
+        Args:
+            app_config (AppConfig): The application configuration object.
+            companion (firestore.DocumentSnapshot): Firestore document snapshot containing companion settings.
+        """
+        # Retrieve settings and use defaults if necessary
+        settings = {
+            "chat_model": (
+                safely_get_field(
+                    companion,
+                    "chat_model",
+                    self.DEFAULT_CONFIG["settings"]["chat_model"],
+                )
+            ),
+            "system_prompt": (
+                safely_get_field(
+                    companion,
+                    "system_prompt",
+                    self.DEFAULT_CONFIG["settings"]["system_prompt"],
+                )
+            ),
+            "temperature": (
+                safely_get_field(
+                    companion,
+                    "temperature",
+                    self.DEFAULT_CONFIG["settings"]["temperature"],
+                )
+            ),
+            "vision_enabled": (
+                safely_get_field(
+                    companion,
+                    "vision_enabled",
+                    self.DEFAULT_CONFIG["settings"]["vision_enabled"],
+                )
+            ),
+        }
+
+        # Add 'prefix_messages_content' only if it exists
+        prefix_messages_content = safely_get_field(companion, "prefix_messages_content")
+        if prefix_messages_content is not None:
+            settings["prefix_messages_content"] = json.dumps(prefix_messages_content)
+
+        # Apply the new configuration settings
+        self.config.read_dict({"settings": settings})
+
     @abstractmethod
     def load_config(self):
         """
@@ -95,3 +149,24 @@ class AppConfig(ABC):
             f"Temperature: {self.config.get('settings', 'temperature')}"
         )
         return readable_config
+
+
+def init_chat_model(app_config: AppConfig) -> ChatOpenAI:
+    """
+    Initialize the langchain chat model.
+
+    Args:
+        app_config (AppConfig): Application configuration object.
+
+    Returns:
+        ChatOpenAI: Initialized chat model.
+    """
+    config = app_config.config
+    chat = ChatOpenAI(
+        model=config.get("settings", "chat_model"),
+        temperature=float(config.get("settings", "temperature")),
+        openai_api_key=config.get("api", "openai_api_key"),
+        openai_organization=config.get("api", "openai_organization", fallback=None),
+        max_tokens=MAX_TOKENS,
+    )  # type: ignore
+    return chat
